requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.18.3
python-telegram-bot==20.0
python-dotenv==1.0.0
